// src/app/shared/types/api-types/bidt-df-type-operation-dto.interface.ts
// //////////////////////////////////////////////////////////////////////////
// DO NOT EDIT THIS FILE
//
// It is automatically generated by generate-dtos.js
// //////////////////////////////////////////////////////////////////////////

import { BidtDfTypeOperationIdDTO } from './bidt-df-type-operation-id-dto.interface';

export interface BidtDfTypeOperationDTO extends BidtDfTypeOperationIdDTO {
  bidtDfTypeId: number;
  dftopeKey?: string; // 50
  dftopeSequenceNumber?: string;
  objectType?: string; // 50
  taskTypeCode?: string; // 50
  pnTypes?: string[];
  docCategories?: string[];
}